chore(test): Move common fixtures and functions to global conftest.py
