+++ tests/CLI_test/conftest.py
@@ -1,91 +1,21 @@
+""" Pytest configuration for CLI tests. """
+
 import os
 from collections.abc import Generator
 from pathlib import Path
 
 import pytest
-from _pytest.logging import LogCaptureFixture
-from loguru import logger
 
-from FABulous.FABulous_CLI.FABulous_CLI import FABulous_CLI
-from FABulous.FABulous_CLI.helper import create_project, setup_logger
-
-
-def normalize(block: str) -> list[str]:
-    """Normalize a block of text to perform comparison.
-
-    Strip newlines from the very beginning and very end, then split into separate lines and strip trailing whitespace
-    from each line.
-    """
-    assert isinstance(block, str)
-    block = block.strip("\n")
-    return [line.rstrip() for line in block.splitlines()]
-
-
-def run_cmd(app: FABulous_CLI, cmd: str) -> None:
-    """Clear stdout, stdin and stderr buffers, run the command, and return stdout and stderr"""
-    app.onecmd_plus_hooks(cmd)
-
-
-def normalize_and_check_for_errors(caplog_text: str) -> list[str]:
-    """Normalize a block of text and check for errors."""
-    log = normalize(caplog_text)
-    assert not any("ERROR" in line for line in log), "Error found in log messages"
-    return log
-
+from FABulous.FABulous_CLI.helper import create_project
 
 TILE = "LUT4AB"
 
 
-@pytest.fixture(autouse=True)
-def env() -> Generator[None]:
-    fabulousRoot = str(Path(__file__).resolve().parent.parent.parent / "FABulous")
-    os.environ["FAB_ROOT"] = fabulousRoot
-    os.environ["FABULOUS_TESTING"] = "TRUE"
-    yield
-    os.environ.pop("FAB_ROOT", None)
-    os.environ.pop("FABULOUS_TESTING", None)
-
-
-@pytest.fixture
-def cli(tmp_path: Path) -> Generator[FABulous_CLI]:
-    projectDir = tmp_path / "test_project"
-    os.environ["FAB_PROJ_DIR"] = str(projectDir)
-    create_project(projectDir)
-    setup_logger(0, False)
-    cli = FABulous_CLI(writerType="verilog", projectDir=projectDir, enteringDir=tmp_path)
-    cli.debug = True
-    run_cmd(cli, "load_fabric")
-    yield cli
-    os.environ.pop("FAB_ROOT", None)
-    os.environ.pop("FAB_PROJ_DIR", None)
-
-
 @pytest.fixture
 def project(tmp_path: Path) -> Generator[Path]:
+    """Create a temporary FABulous project directory."""
     project_dir = tmp_path / "test_project"
     os.environ["FAB_PROJ_DIR"] = str(project_dir)
     create_project(project_dir)
     yield project_dir
     os.environ.pop("FAB_PROJ_DIR", None)
-
-
-@pytest.fixture(autouse=True)
-def cleanup_logger() -> Generator[None]:
-    """Ensure logger is properly cleaned up after each test to prevent
-    'logging to closed file' errors when tests exit quickly"""
-    yield
-    # Remove all logger handlers to prevent logging to closed files
-    logger.remove()
-
-
-@pytest.fixture
-def caplog(caplog: LogCaptureFixture) -> Generator[LogCaptureFixture]:
-    handler_id = logger.add(
-        caplog.handler,
-        format="{message}",
-        level=0,
-        filter=lambda record: record["level"].no >= caplog.handler.level,
-        enqueue=False,  # Set to 'True' if your test is spawning child processes.
-    )
-    yield caplog
-    logger.remove(handler_id)
