+++ tests/CLI_test/test_CLI.py
@@ -4,7 +4,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from FABulous.FABulous_CLI.FABulous_CLI import FABulous_CLI
-from tests.CLI_test.conftest import (
+from tests.conftest import (
     TILE,
     normalize_and_check_for_errors,
     run_cmd,
